pyproject.toml
[tool.poetry]
name = "healthy-habit-tracker"
version = "0.1.0"
description = ""
authors = ["Andrew_Polik <polik.24@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
python-decouple = "^3.8"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.32.3"
celery = "^5.4.0"
redis = "^5.0.4"
eventlet = "^0.36.1"
django-celery-beat = "^2.6.0"
django-cors-headers = "^4.3.1"
drf-yasg = "^1.21.7"
coverage = "^7.5.3"
flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
